requirements.txt
PyYAML>=6.0
markdown-pdf
